@@ -4,7 +4,7 @@ description: Demonstrates how to use the adhoc_plugin plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
