@@ -3,8 +3,8 @@ abstract class AodvMessage {
   late int _type;
 
   /// Creates an [AodvMessage] object.
-  /// 
-  /// The type of message is specified by [type]. 
+  ///
+  /// The type of message is specified by [type].
   AodvMessage(int type) {
     _type = type;
   }
