@@ -39,39 +39,52 @@ class GroupController {
 
   /// Order of the finite cyclic group
   BigInt? _p;
+
   /// Generator of the finite cyclic group of order [_p]
   BigInt? _g;
+
   /// Private Diffie-Hellman share
   BigInt? _x;
+
   /// Private key share
   BigInt? _k;
+
   /// Private share
   BigInt? _d;
+
   /// Secret group key
   SecretKey? _groupKey;
+
+  /// Secret group key hash
+  BigInt? _groupKeyHash;
+
   /// Time allowed for joining the group creation process
   late int _expiryTime;
+
   /// Group member's key share recovered
   late int _recovered;
+
   /// Label of the group initiator/owner
   late String? _groupOwner;
+
   /// Map containing the Diffie-Hellman share of each member
   late HashMap<String, BigInt> _DHShare;
+
   /// Map containing the member share of each member
   late HashMap<String, BigInt> _memberShare;
+
   /// Map containing the Chinese Remainder Theorem solution of each member
   late HashMap<String, BigInt> _CRTShare;
+
   /// List containing the group member label
   late List<String> _memberLabel;
 
   /// Creates a [GroupController] object.
-  /// 
-  /// This object is configured according to [config], which contains specific 
+  ///
+  /// This object is configured according to [config], which contains specific
   /// configurations.
-  GroupController(
-    this._engine, this._aodvManager, this._datalinkManager, this._eventStream, 
-    Config config
-  ) {
+  GroupController(this._engine, this._aodvManager, this._datalinkManager,
+      this._eventStream, Config config) {
     _ownLabel = _aodvManager.label;
     _controller = StreamController<AdHocEvent>.broadcast();
     _open = config.public;
@@ -101,7 +114,7 @@ class GroupController {
 /*-------------------------------Public methods-------------------------------*/
 
   /// Initiates a secure group creation process
-  /// 
+  ///
   /// If [members] is given, then the group formation request is send to every
   /// member in the list.
   void createGroup([List<String>? members]) {
@@ -116,14 +129,15 @@ class GroupController {
 
     // Timestamp for flood control
     var timestamp = _ownLabel + DateTime.now().toIso8601String();
-    var low = 32;
-    var seed = max(low + low, Random(42).nextInt(192));
+    var low = 8;
+    var seed = max(low + low, Random(42).nextInt(32));
 
     // Choose Diffie-Hellman parameters
     _p = randomPrimeBigInt(seed);
     _g = randomPrimeBigInt(low);
 
-    var info = GroupInit(timestamp, _p.toString(), _g.toString(), _ownLabel, false);
+    var info =
+        GroupInit(timestamp, _p.toString(), _g.toString(), _ownLabel, false);
     var message = SecureData(GROUP_INIT, info.toJson());
 
     if (members == null) {
@@ -140,9 +154,8 @@ class GroupController {
     Timer(Duration(seconds: _expiryTime), _broadcastTimerExpired);
   }
 
-
   /// Joins an existing secure group
-  /// 
+  ///
   /// If [label] is given, then the join group request is sent to label
   void joinSecureGroup([String? label]) {
     // Send a group join request
@@ -157,10 +170,9 @@ class GroupController {
     }
   }
 
-
   /// Leaves an existing secure group
   void leaveSecureGroup() {
-    if (!_isFormationGoingOn || !_isGroupFormed) {
+    if (_isFormationGoingOn || !_isGroupFormed) {
       return;
     }
 
@@ -180,11 +192,10 @@ class GroupController {
     _isFormationGoingOn = false;
   }
 
-
   /// Sends a encrypted message to the secure group.
-  /// 
+  ///
   /// The message payload is set to [data] and is encrypted using the group key.
-  /// 
+  ///
   /// Throws a [GroupNotFormedException] exception if the group is not formed
   /// or the device is not part of any secure group.
   void sendMessageToGroup(Object? data) async {
@@ -193,7 +204,7 @@ class GroupController {
     }
 
     var encrypted = await _engine.encrypt(
-      Utf8Encoder().convert(JsonCodec().encode(data)), 
+      Utf8Encoder().convert(JsonCodec().encode(data)),
       sharedKey: _groupKey!,
     );
 
@@ -214,23 +225,20 @@ class GroupController {
       if (event.type == DATA_RECEIVED) {
         var sender = (event.payload as List<dynamic>)[0] as AdHocDevice;
         var payload = SecureData.fromJson(
-          (event.payload as List<dynamic>)[1] as Map<String, dynamic>
-        );
+            (event.payload as List<dynamic>)[1] as Map<String, dynamic>);
 
         _processDataReceived(sender, payload);
       }
     });
   }
 
-
   /// Computes a Diffie-Hellman share.
-  /// 
+  ///
   /// Returns a integer value as [BigInt] representing y = g^x mod p.
   BigInt _computeDHShare() {
     _x = randomBigInt(_p!.bitLength, max: _p);
     return _g!.modPow(_x!, _p!);
   }
-
 
   /// Triggers the start of the group key agreement.
   void _broadcastTimerExpired() {
@@ -256,29 +264,27 @@ class GroupController {
     }
   }
 
-  
   /// Computes the Diffie-Hellman secret share.
-  /// 
+  ///
   /// The Diffie-Hellman secret share is computed from [yj] sent by a remote
   /// node identified by [label].
-  /// 
+  ///
   /// Returns an integer value as [BigInt] representing the secret share.
   BigInt _computeMemberShare(String label, BigInt yj) {
     var mij = yj.modPow(_x!, _p!);
-    mij = mij > (_p!~/BigInt.two) ? mij : _p! - mij;
+    mij = mij > (_p! ~/ BigInt.two) ? mij : _p! - mij;
     _memberShare.putIfAbsent(label, () => mij);
     return mij;
   }
 
-  
   /// Solves the CRT system of congruences.
-  /// 
+  ///
   /// The CRT system is solved with the parameters sent by a remote node [label].
-  /// 
-  /// The public D-H share and secret shared D-H share is set respectively to 
+  ///
+  /// The public D-H share and secret shared D-H share is set respectively to
   /// [yj] and [mij].
-  /// 
-  /// Returns the solution as an integer [BigInt] to the CRT system of 
+  ///
+  /// Returns the solution as an integer [BigInt] to the CRT system of
   /// congruences.
   BigInt _computeCRTShare(String label, BigInt yj, BigInt mij) {
     BigInt pij, _min = _memberShare.values.first;
@@ -313,22 +319,21 @@ class GroupController {
       }
     }
 
-    // Solve the system of congruences (Chinese Remainder Theorem) using the 
+    // Solve the system of congruences (Chinese Remainder Theorem) using the
     // Bézout's identity to obtain crt_ij
     var coefficients = _solveBezoutIdentity(mij, pij);
     var crtij = (_k! * coefficients[1] * pij) + (_d! * coefficients[0] * mij);
     if (crtij < BigInt.zero) {
-      crtij % (mij * pij);
+      crtij = crtij % (mij * pij);
     }
 
     return crtij;
   }
 
-
   /// Solves the Bézout identity.
-  /// 
+  ///
   /// The system parameters is set to [a] and [b].
-  /// 
+  ///
   /// Returns a list of integer value as [BigInt] that represents the solution
   /// of the system.
   List<BigInt> _solveBezoutIdentity(BigInt a, BigInt b) {
@@ -336,33 +341,33 @@ class GroupController {
     var V = BigInt.zero, _V = BigInt.one;
 
     while (_R != BigInt.zero) {
-      var Q = R~/_R;
+      var Q = R ~/ _R;
       var RS = R, US = U, VS = V;
-      R = _R; U = _U; V = _V;
-      _R = RS - Q*_R;
-      _U = US - Q*_U;
-      _V = VS - Q*_V;
+      R = _R;
+      U = _U;
+      V = _V;
+      _R = RS - Q * _R;
+      _U = US - Q * _U;
+      _V = VS - Q * _V;
     }
 
     return List.empty(growable: true)..add(U)..add(V);
   }
 
-  
   /// Gets the bytes representation of a [BigInt].
-  /// 
+  ///
   /// The given big int value to represent is given by [bigInt].
-  /// 
+  ///
   /// Returns the bytes representation of the given value.
   Uint8List _toBytes(BigInt bigInt) {
     const BYTE_SIZE = 8;
 
-    var byteData = ByteData((bigInt.bitLength~/BYTE_SIZE) + 1);
+    var byteData = ByteData((bigInt.bitLength ~/ BYTE_SIZE) + 1);
     var _bigInt = bigInt;
 
     for (var i = 1; i <= byteData.lengthInBytes; i++) {
       byteData.setUint8(
-        byteData.lengthInBytes - i, _bigInt.toUnsigned(BYTE_SIZE).toInt()
-      );
+          byteData.lengthInBytes - i, _bigInt.toUnsigned(BYTE_SIZE).toInt());
 
       _bigInt = _bigInt >> BYTE_SIZE;
     }
@@ -370,36 +375,38 @@ class GroupController {
     return byteData.buffer.asUint8List();
   }
 
-  
   /// Computes a hash of the group key.
   Future<BigInt> _computeGroupKeyHash() async {
     final hash = await Sha256().hash(await _groupKey!.extractBytes());
     return BigInt.from(hash.bytes.reduce((a, b) => a + b));
   }
 
-
   /// Computes the group key.
-  /// 
+  ///
   /// The way the group key is computed is defined by [type].
   void _computeGroupKey(int type, [String? label]) async {
     var groupKeySum = BigInt.zero;
-    var groupKeyHash = BigInt.zero;
 
     switch (type) {
       case GROUP_INIT:
         groupKeySum += _k!;
         for (final label in _CRTShare.keys) {
-          groupKeySum ^= (_CRTShare[label]! % _memberShare[label]!);
+          if (label != _ownLabel) {
+            groupKeySum ^= (_CRTShare[label]! % _memberShare[label]!);
+          }
         }
         break;
 
       case GROUP_JOIN:
-        groupKeyHash = await _computeGroupKeyHash();
+        if (_groupKeyHash == null) {
+          _groupKeyHash = await _computeGroupKeyHash();
+        }
 
         if (label == null) {
-          groupKeySum = groupKeyHash ^ _k!;
+          groupKeySum = _groupKeyHash! ^ _k!;
         } else {
-          groupKeySum = groupKeyHash ^ (_CRTShare[label]! % _memberShare[label]!);
+          groupKeySum =
+              _groupKeyHash! ^ (_CRTShare[label]! % _memberShare[label]!);
         }
         break;
 
@@ -418,7 +425,8 @@ class GroupController {
     var keyBytes = _toBytes(groupKeySum);
     var length = keyBytes.length;
     if (length < keyLengthRequired) {
-      keyBytes = Uint8List.fromList(keyBytes.toList() + List.filled(keyLengthRequired - length, 42));
+      keyBytes = Uint8List.fromList(keyBytes.toList().sublist(0, length) +
+          List.filled(keyLengthRequired - length, 42));
     } else if (length > keyLengthRequired) {
       keyBytes = keyBytes.sublist(0, keyLengthRequired);
     }
@@ -428,15 +436,15 @@ class GroupController {
 
     if (!_isGroupFormed) {
       _isGroupFormed = true;
-      _controller.add(AdHocEvent(GROUP_STATUS, type));
     }
 
     _isFormationGoingOn = false;
+
+    _controller.add(AdHocEvent(GROUP_STATUS, type));
   }
 
-
   /// Processes the data received.
-  /// 
+  ///
   /// The data is retrieved from the [event] payload.
   void _processDataReceived(AdHocDevice sender, SecureData secureData) async {
     var senderLabel = sender.label!;
@@ -444,7 +452,8 @@ class GroupController {
     switch (secureData.type) {
       case GROUP_INIT:
         // Retrieve group advertisement data
-        var data = GroupInit.fromJson(secureData.payload as Map<String, dynamic>);
+        var data =
+            GroupInit.fromJson(secureData.payload as Map<String, dynamic>);
 
         // Advertisement flood control
         if (!_setFloodEvents.contains(data.timestamp)) {
@@ -476,7 +485,6 @@ class GroupController {
         _aodvManager.sendMessageTo(senderLabel, msg);
         break;
 
-
       case GROUP_REPLY:
         if (_timerExpired) {
           return;
@@ -487,10 +495,10 @@ class GroupController {
         }
         break;
 
-
       case GROUP_LIST:
         // Retrieve group list data
-        var data = GroupList.fromJson(secureData.payload as Map<String, dynamic>);
+        var data =
+            GroupList.fromJson(secureData.payload as Map<String, dynamic>);
         // Get all the label of the group member
         _memberLabel.addAll(data.labels);
 
@@ -507,10 +515,10 @@ class GroupController {
         }
         break;
 
-
       case GROUP_SHARE:
         // Retrieve group share data
-        var data = GroupValue.fromJson(secureData.payload as Map<String, dynamic>);
+        var data =
+            GroupValue.fromJson(secureData.payload as Map<String, dynamic>);
         // Store the public Diffie-Hellman share of group memeber
         var yj = BigInt.parse(data.value);
         _DHShare.putIfAbsent(senderLabel, () => yj);
@@ -527,20 +535,24 @@ class GroupController {
 
             var mij = _computeMemberShare(label, yj);
             var crtij = _computeCRTShare(label, yj, mij);
-            var msg = SecureData(GROUP_KEY, GroupValue(crtij.toString()).toJson());
+            var msg =
+                SecureData(GROUP_KEY, GroupValue(crtij.toString()).toJson());
 
             _aodvManager.sendMessageTo(label, msg.toJson());
           }
         }
         break;
 
-
       case GROUP_KEY:
         // Retrieve group key share data
-        var data = GroupValue.fromJson(secureData.payload as Map<String, dynamic>);
+        var data =
+            GroupValue.fromJson(secureData.payload as Map<String, dynamic>);
         var crtji = BigInt.parse(data.value);
         // Store the solution of the CRT system of congruence
-        _CRTShare.putIfAbsent(senderLabel, () { _recovered += 1; return crtji; });
+        _CRTShare.putIfAbsent(senderLabel, () {
+          _recovered += 1;
+          return crtji;
+        });
 
         // Compute the group key
         if (_recovered == _memberLabel.length) {
@@ -548,17 +560,16 @@ class GroupController {
         }
         break;
 
-
       case GROUP_JOIN:
         // Send the group join request to the group owner
         var msg = SecureData(GROUP_JOIN_REQ, GroupValue(senderLabel).toJson());
         _aodvManager.sendMessageTo(_groupOwner!, msg);
         break;
 
-
       case GROUP_JOIN_REQ:
         // Retrieve joining member label data
-        var data = GroupValue.fromJson(secureData.payload as Map<String, dynamic>);
+        var data =
+            GroupValue.fromJson(secureData.payload as Map<String, dynamic>);
         var joiningMember = data.value;
         if (_memberLabel.contains(joiningMember)) {
           break;
@@ -577,32 +588,40 @@ class GroupController {
         });
 
         var response = GroupJoin(
-          hash: groupKeyHash.toString(), labels: labels, values: values
-        );
+            hash: groupKeyHash.toString(),
+            labels: labels,
+            values: values,
+            share: _p.toString(),
+            solution: _g.toString());
 
         var msg = SecureData(GROUP_JOIN_REP, response.toJson());
 
         _aodvManager.sendMessageTo(joiningMember, msg.toJson());
         break;
 
-
       case GROUP_JOIN_REP:
         // Retrieve joining member label data
-        var data = GroupJoin.fromJson(secureData.payload as Map<String, dynamic>);
+        var data =
+            GroupJoin.fromJson(secureData.payload as Map<String, dynamic>);
 
         // New member proceeds with the protocol from Step 3.
-        if (data.hash != null) {
+        if (data.hash != 'null') {
+          _groupOwner = senderLabel;
+
+          _p = BigInt.parse(data.share!);
+          _g = BigInt.parse(data.solution!);
+
           // Hash of group key
-          _groupKey = SecretKey(_toBytes(BigInt.parse(data.hash!)));
+          _groupKeyHash = BigInt.parse(data.hash!);
 
           // Public Diffie-Hellman shares of group members
           var labels = data.labels;
           var values = data.values;
           var y = _computeDHShare();
 
-          // Broadcast own Diffie-Hellman public share
+          _memberLabel.addAll(labels!);
           _DHShare.putIfAbsent(_ownLabel, () => y);
-          for (var i = 0; i < labels!.length; i++) {
+          for (var i = 0; i < labels.length; i++) {
             _DHShare.putIfAbsent(labels[i], () => BigInt.parse(values![i]));
           }
 
@@ -612,7 +631,10 @@ class GroupController {
               var mij = _computeMemberShare(label, _DHShare[label]!);
               var crtij = _computeCRTShare(label, _DHShare[label]!, mij);
 
-              var response = GroupJoin(share: y.toString(), solution: crtij.toString());
+              var response = GroupJoin(
+                  hash: 'null',
+                  share: y.toString(),
+                  solution: crtij.toString());
               var msg = SecureData(GROUP_JOIN_REP, response.toJson());
 
               _aodvManager.sendMessageTo(label, msg);
@@ -620,7 +642,10 @@ class GroupController {
           }
           // Compute group key
           _computeGroupKey(GROUP_JOIN, null);
-        } else { // Old member updating the group key
+        } else {
+          // Old member updating the group key
+          _memberLabel.add(senderLabel);
+
           var yj = BigInt.parse(data.share!);
           var mij = _computeMemberShare(senderLabel, yj);
           var crtij = BigInt.parse(data.solution!);
@@ -635,10 +660,10 @@ class GroupController {
         }
         break;
 
-
       case GROUP_LEAVE:
         // Retrieve joining member label data
-        var data = GroupLeave.fromJson(secureData.payload as Map<String, dynamic>);
+        var data =
+            GroupLeave.fromJson(secureData.payload as Map<String, dynamic>);
 
         // Group owner redraw new key share and broadcast it to remaining group member
         if (_groupOwner == _ownLabel) {
@@ -652,14 +677,17 @@ class GroupController {
               // Redraw new key share
               var crtij = _computeCRTShare(label, yj, mij);
 
-              var response = GroupLeave(data.leavingLabel, newSolution: crtij.toString());
+              var response =
+                  GroupLeave(data.leavingLabel, newSolution: crtij.toString());
               var msg = SecureData(GROUP_LEAVE, response.toJson());
 
               _aodvManager.sendMessageTo(label, msg);
             }
           }
           // Compute group key
-          _computeGroupKey(GROUP_LEAVE);
+          if (_memberLabel.length > 1) {
+            _computeGroupKey(GROUP_LEAVE);
+          }
         } else {
           // Remove value of leaving member
           var leavingMember = data.leavingLabel;
@@ -674,10 +702,11 @@ class GroupController {
         }
         break;
 
-
       case GROUP_DATA:
         // Decrypt group data received
-        var decrypted = await _engine.decrypt(secureData.payload as List<dynamic>, sharedKey: _groupKey!);
+        var decrypted = await _engine.decrypt(
+            secureData.payload as List<dynamic>,
+            sharedKey: _groupKey!);
         var processed = JsonCodec().decode(Utf8Decoder().convert(decrypted));
 
         // Notify upper layers of group data received
