@@ -67,10 +67,10 @@ class CertificateRepository {
   ///
   /// If a certificate has expired, then it is simply removed.
   void _checkCertificatesValidity() {
-    Timer.periodic(Duration(seconds: _period),
-      (timer) => _repository.removeWhere(
-        (label, certificate) => certificate.validity.isBefore(DateTime.now())
-      ),
+    Timer.periodic(
+      Duration(seconds: _period),
+      (timer) => _repository.removeWhere((label, certificate) =>
+          certificate.validity.isBefore(DateTime.now())),
     );
   }
 }
